FEATURE: Add node tags
Each node has a tags attribute, which is a set of strings. Methods set_tag, set_tags, clear_tag and clear_tags are provided to manipulate tags. Tags can also be set as a parameter to add_node. A tags method allows access, as well as an attribute-based accessor 't', in the style of 'v' for values and 's' for state.

Strings of the form '__[name]__' are reserved for internal Loman use.

Serialization and Expansion nodes are now represented using tags.